refactor: struct -> ws
