+++ packages/ws/src/index.ts
@@ -1,7 +1,14 @@
-export * from './strategies/sharding/IShardingStrategy';
 export * from './strategies/context/IContextFetchingStrategy';
+export * from './strategies/context/SimpleContextFetchingStrategy';
+export * from './strategies/context/WorkerContextFetchingStrategy';
+
+export * from './strategies/sharding/IShardingStrategy';
 export * from './strategies/sharding/SimpleShardingStrategy';
-export * from './struct/WebSocketManager';
-export * from './struct/WebSocketShard';
+export * from './strategies/sharding/WorkerShardingStrategy';
+
 export * from './utils/constants';
+export * from './utils/IdentifyThrottler';
 export * from './utils/utils';
+
+export * from './ws/WebSocketManager';
+export * from './ws/WebSocketShard';
